% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_decade_table.R
\name{build_decade_table}
\alias{build_decade_table}
\alias{create_base_table}
\alias{set_base_options}
\alias{add_format_style}
\title{Build a trend table}
\usage{
build_decade_table(data_object)

create_base_table(data_object)

set_base_options()

add_format_style(table)
}
\arguments{
\item{data_object}{A data object - either a matrix, data frame or data table.}

\item{table}{A \code{datatables} class object created using \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
An HTML Widget object containing \code{datatables} table. Please check
\href{https://rstudio.github.io/DT/}{DT} for more details.
}
\description{
Generates a \code{datatables} class table to show trend data aggregated to decades
for the selected indicator and countries.
}
\details{
Utilizes collection of functions trough the pipeline. Each "add_" function
appends relevant layer do the table object. Each "set_" function is
responsible for relevant configuration options.
}
