% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{build_trend_chart}
\alias{build_trend_chart}
\alias{create_base_chart}
\alias{get_series_args}
\alias{add_series}
\alias{add_layout}
\alias{add_config}
\alias{add_event_register}
\alias{set_plotly_remove_btns}
\title{Build a trends chart}
\usage{
build_trend_chart(dt_list, color_theme, layout_config)

create_base_chart()

get_series_args(dt_list, series_color)

add_series(chart, dt_list, series_color)

add_layout(chart, layout_config, legend_color)

add_config(chart)

add_event_register(chart)

set_plotly_remove_btns()
}
\arguments{
\item{dt_list}{A named list with a data table (see Details).}

\item{color_theme}{A named list of color themes (see Details).}

\item{layout_config}{A named list of layout configuration (see Details).}

\item{series_color}{A string with a line and marker color for
the trend chart series.}

\item{chart}{A \code{plotly} class object created using \code{\link[plotly:plot_ly]{plotly::plot_ly()}}.}

\item{legend_color}{A string with a background and border color for
the trend chart legend.}
}
\value{
An HTML Widget containing \code{plotly} chart. Please check
\href{https://www.rdocumentation.org/packages/plotly/topics/plot_ly}{plotly}
for more details.
}
\description{
Generates a \code{plotly} class chart to show trend data for the selected
indicator and countries.
}
\details{
Utilizes collection of functions trough the pipeline. Each "add_" function
appends relevant layer do the trend chart object. Each "set_" function is
responsible for relevant configuration options.

\strong{\code{dt_list}} list needs to contain only one element - a data table with
a values to plot. The name of the element is a country label (name).

\strong{\code{color_theme}} list needs to contain two elements - "series"
and "legend". "series" is a color for the trend chart series line
and marker. "legend" is a color for the trend chart legend background
and border.

\strong{\code{layout_config}} list needs to contain two elements - "title"
and "yrange". "title" is the trend chart title (and data indicator name).
"yrange" is a pair of integers - initial real (no extra padding) minimum
and maximum values of y-axis.
}
