% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_world_map.R
\name{build_world_map}
\alias{build_world_map}
\alias{create_base_map}
\alias{add_bounds}
\alias{add_view}
\alias{add_clickable_polygons}
\alias{add_notclickable_polygons}
\alias{set_label_options}
\alias{set_highlight_options}
\alias{add_reset_button}
\title{Build a world map}
\usage{
build_world_map(data_list, color_theme)

create_base_map()

add_bounds(map)

add_view(map)

add_clickable_polygons(map, sf_data, color_theme)

add_notclickable_polygons(map, sf_data, color_theme)

set_label_options()

set_highlight_options(stroke_color)

add_reset_button(map)
}
\arguments{
\item{data_list}{A named list of \code{sf}data frames (see Details).}

\item{color_theme}{A named list of color themes (see Details).}

\item{map}{A \code{leaflet} class object created using \code{\link[leaflet:leaflet]{leaflet::leaflet()}}.}

\item{sf_data}{An \code{sf} data frame (see Details).}

\item{stroke_color}{A string with a stroke color for polygon highlights.}
}
\value{
An HTML Widget object containing \code{leaflet} map. Please check
\href{https://rstudio.github.io/leaflet/reference/leaflet.html}{leaflet}
for more details.
}
\description{
Generates a clickable \code{leaflet} class world map to use with the trends
data.
}
\details{
Utilizes collection of functions trough the pipeline. Each "add_" function
appends relevant layer do the map object. Each "set_" function is responsible
for polygon options configuration.

\strong{\code{data_list}} list needs to contain at least two elements - "clickable"
and "notclickable". Each of this elements is expected to be a data frame with
exact three columns:
\itemize{
\item "id" - integer, ISO 3166-1 numeric (or numeric-3) country codes.
\item "label" - character, country names.
\item "geometry" - \code{sfc} (simple feature collection) class object of MULTIPOLYGON
country geometries.
}

\strong{\code{color_theme}} list needs to contain at least two elements - "clickable"
and "notclickable". Each of this elements is expected to be a list with
color definition for polygon "stroke", "fill" and "highlight_stroke".
}
