% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{convert_dt_to_string}
\alias{convert_dt_to_string}
\alias{convert_dt_to_json}
\title{Convert a data table to character string or JSON format}
\usage{
convert_dt_to_string(data_object)

convert_dt_to_json(data_object)
}
\arguments{
\item{data_object}{An object that can be coerced to \code{data.table}.}
}
\value{
A \code{character} string or \code{json} object.
}
\description{
Converts a data table to character string or JSON format.
}
