% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{create_chart_config}
\alias{create_chart_config}
\title{Create chart configuration}
\usage{
create_chart_config(
  indic_id,
  initial_range,
  indic_name_id = set_name_id_indicators()
)
}
\arguments{
\item{indic_id}{A string with a data indicator id.}

\item{initial_range}{A vector containing a pair of integers where
the first element is the data range real minimum
and the second is the real maximum.}

\item{indic_name_id}{A list where the elements are data indicator ids
and the element names are indicator "human friendly" names. Default
values are provided by \code{\link[=set_name_id_indicators]{set_name_id_indicators()}}.}
}
\value{
A named list with a configuration for the trend chart layout.
}
\description{
Makes a configuration list to use in the trends chart layout.
}
