% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_copyright_note}
\alias{create_copyright_note}
\title{Create copyright footnote}
\usage{
create_copyright_note()
}
\value{
A string with a copyright note.
}
\description{
Creates copyright footnote to use across the app.
}
