% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_data_extractor.R
\name{create_db_connection}
\alias{create_db_connection}
\title{Create a connection to MySQL database}
\usage{
create_db_connection(secrets)
}
\arguments{
\item{secrets}{A named character vector that contains a configuration data
for a MySQL database connection.}
}
\value{
A connection to a database. Please check \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
for more details.
}
\description{
Creates a connection to relevant MySQL database using provided secrets.
}
