% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shiny_tags.R
\name{create_header_hlink}
\alias{create_header_hlink}
\alias{create_example_hlink}
\alias{create_footer_hlink}
\title{Create a hyperlink for header/example/footer division}
\usage{
create_header_hlink(hlink_label)

create_example_hlink(hlink_label)

create_footer_hlink(hlink_label, hlink_ref)
}
\arguments{
\item{hlink_label}{A string with a hyperlink label.}

\item{hlink_ref}{A string with the URL link to the external web page.
Used only in \code{\link[=create_footer_hlink]{create_footer_hlink()}} function.}
}
\value{
A single \code{shiny.tag} object.
}
\description{
Constructs a hyperlink (\if{html}{\out{<a>}}\if{html}{\out{</a>}}) to use in the main page header,
example or footer section.
}
