% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_range.R
\name{create_xrange}
\alias{create_xrange}
\alias{create_yrange}
\alias{calculate_yrange_padding}
\title{Create range for data on x or y axis}
\usage{
create_xrange(initial_range)

create_yrange(initial_range)

calculate_yrange_padding(initial_range)
}
\arguments{
\item{initial_range}{A vector containing a pair of integers where
the first element is the data range observed minimum
and the second is the observed maximum.}
}
\value{
A vector containing a pair of integers where the first element is
the data range modified minimum and the second is the modified maximum.
}
\description{
Creates x or y axis data range for \code{plotly} line chart.
}
\details{
For x-axis new range is created using \code{celinig()} and \code{floor()} rounding.

For y-axis extra padding is added around its initial minimum and maximum
values.
The goal is to avoid layout elements (e.g. lines, markers, tick labels),
that are near the edges of the plotting area, being clipped or cut off.
}
