% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_data_extractor.R
\name{DbDataExtractor}
\alias{DbDataExtractor}
\title{Database data extractor}
\description{
Fetches data from MySQL database.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{db_schema}}{A named list with SQL table names as the list element
names and column names, in a qualified column name format, as the list
elements.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{query}}{A read only list of SQL queries created inside the class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DbDataExtractor-new}{\code{DbDataExtractor$new()}}
\item \href{#method-DbDataExtractor-get_data}{\code{DbDataExtractor$get_data()}}
\item \href{#method-DbDataExtractor-get_geometry_clickable}{\code{DbDataExtractor$get_geometry_clickable()}}
\item \href{#method-DbDataExtractor-get_geometry_notclickable}{\code{DbDataExtractor$get_geometry_notclickable()}}
\item \href{#method-DbDataExtractor-get_indicator_data}{\code{DbDataExtractor$get_indicator_data()}}
\item \href{#method-DbDataExtractor-get_weight_data}{\code{DbDataExtractor$get_weight_data()}}
\item \href{#method-DbDataExtractor-clear_result}{\code{DbDataExtractor$clear_result()}}
\item \href{#method-DbDataExtractor-disconnect}{\code{DbDataExtractor$disconnect()}}
\item \href{#method-DbDataExtractor-clone}{\code{DbDataExtractor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-new"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-new}{}}}
\subsection{Method \code{new()}}{
Creates a new extractor object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$new(con, db_schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{A connection to database. Please check \code{\link[DBI:dbConnect]{DBI::dbConnect()}}
for more details.}

\item{\code{db_schema}}{A list with database schema.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DbDataExtractor} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-get_data}{}}}
\subsection{Method \code{get_data()}}{
Fetches relevant data from the target database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$get_data(type, country_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string indicating type of the data to fetch.}

\item{\code{country_id}}{An integer, ISO 3166-1 numeric (or numeric-3) country.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-get_geometry_clickable"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-get_geometry_clickable}{}}}
\subsection{Method \code{get_geometry_clickable()}}{
Fetches "clickable" geometry data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$get_geometry_clickable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-get_geometry_notclickable"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-get_geometry_notclickable}{}}}
\subsection{Method \code{get_geometry_notclickable()}}{
Fetches "notclickable" geometry data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$get_geometry_notclickable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-get_indicator_data"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-get_indicator_data}{}}}
\subsection{Method \code{get_indicator_data()}}{
Fetches data for target indicator in target country.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$get_indicator_data(country_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country_id}}{An integer, ISO 3166-1 numeric (or numeric-3) country
code.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-get_weight_data"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-get_weight_data}{}}}
\subsection{Method \code{get_weight_data()}}{
Fetches data for target weight in target country.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$get_weight_data(country_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{country_id}}{An integer, ISO 3166-1 numeric (or numeric-3) country
code.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-clear_result"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-clear_result}{}}}
\subsection{Method \code{clear_result()}}{
Just a wrapper for \code{\link[DBI:dbClearResult]{DBI::dbClearResult()}}. Clears a result set obtained
by the query.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$clear_result(result)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{An object inheriting from \link[DBI:DBIResult-class]{DBI::DBIResult}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Just a wrapper for \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}. Closes connection to the data
base.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DbDataExtractor-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DbDataExtractor-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DbDataExtractor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
