% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_decade_table.R
\name{decadeTableUI}
\alias{decadeTableUI}
\alias{decadeTableServer}
\alias{get_table_description}
\title{Decades table Shiny module}
\usage{
decadeTableUI(id)

decadeTableServer(id, render_flag, aggregation_input)

get_table_description(func_name)
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{render_flag}{A \code{reactiveVal} object containing Boolean value.}

\item{aggregation_input}{A list containing input necessary for the data
aggregation (see Details).}

\item{func_name}{An aggregation function name.}
}
\description{
Provides UI and server logic for a \code{datatables} class decades table.
}
\details{
\strong{\code{aggregation_input}} list needs to contain exactly three \code{reactiveVal}
elements:
\itemize{
\item parameters - a list with "function_name" and "column_name".
\item indicator data - a list of data tables with indicator data per country.
\item weight data - a list that can be empty or contain data tables with data
necessary for weighting per country.
}
}
