% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{extract_download_btn_click}
\alias{extract_download_btn_click}
\title{Extract trends chart download button click}
\usage{
extract_download_btn_click(click_set)
}
\arguments{
\item{click_set}{A list containing two elements. Both are named vectors with
data about clicks. The first element is current click count and the second
one is a snapshot of the previous click count.}
}
\value{
A string (name) identifying clicked button.
}
\description{
Finds out which download button exactly was clicked by comparing current
click info to stored click data.
}
\examples{
new_clicks <- setNames(c(3, 2, 4), c("n1", "n2", "n3"))
old_clicks <- setNames(c(3, 2, 3), c("n1", "n2", "n3"))

extract_download_btn_click(list(new_clicks, old_clicks))
}
