% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{extract_dt_list_range}
\alias{extract_dt_list_range}
\alias{get_dt_range}
\title{Extract a range values from a list of data tables}
\usage{
extract_dt_list_range(data_list, config)

get_dt_range(data_object, col_name)
}
\arguments{
\item{data_list}{A list containing objects that can be coerced to
\code{data.table}.}

\item{config}{A list with a configuration data (see Details).}

\item{data_object}{An object that can be coerced to \code{data.table}.}

\item{col_name}{A character string with a column name for which data table
operation will be conducted.}
}
\value{
A vector containing a pair of numeric values - minimum and maximum.
}
\description{
Finds range values for a given column from a list of data tables.
}
\details{
\strong{\code{config}} list needs to have maximum two elements. First element
is mandatory and contains a column name for which range will be calculated.
Second element is optional and contains filtering details - column name and
condition values.
}
