% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{extract_trends_chart_dt}
\alias{extract_trends_chart_dt}
\title{Extract data that are in the trend chart}
\usage{
extract_trends_chart_dt(data_list, config)
}
\arguments{
\item{data_list}{A list containing objects that can be coerced to
\code{data.table}.}

\item{config}{A list with a configuration data (see Details).}
}
\value{
A data table with the filtered trends chart data in the long format.
}
\description{
Extracts current data that are shown in the trend chart UI.
}
\details{
\strong{\code{config}} list needs to have exact two elements. First element
is a column name which will be used as \code{idcol} in \link[data.table:rbindlist]{data.table::rbindlist}
when creating the long format data. Second element is a filtering details
\itemize{
\item column name and condition values.
}
}
