% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_components.R
\name{find_all_components}
\alias{find_all_components}
\alias{find_ui_components}
\alias{find_server_components}
\alias{find_other_components}
\title{Find all, UI, server or other component files}
\usage{
find_all_components(sub_dir, main_dir)

find_ui_components(sub_dir = set_app_components(), main_dir = "components")

find_server_components(sub_dir = set_app_components(), main_dir = "components")

find_other_components(sub_dir = set_app_components(), main_dir = "components")
}
\arguments{
\item{sub_dir}{A character vector of component sub-directories used in
the app. Default values are provided by \code{\link[=set_app_components]{set_app_components()}}.}

\item{main_dir}{Main component directory. Defaults to \code{"components"}.}
}
\value{
A character vector with component file paths.
}
\description{
Finds app all, UI, server or other (than UI and server) component files
in the relevant directories.
}
