% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Geometric mean}
\usage{
geometric.mean(input_values, na.rm = TRUE)
}
\arguments{
\item{input_values}{A numeric or complex vector.}

\item{na.rm}{A Boolean value. Whether NA values should be stripped before
the computation proceed. Used by \code{\link[base:mean]{base::mean()}}.}
}
\value{
A single numeric value.
}
\description{
Calculates geometric mean using \code{\link[base:Log]{base::log()}}, \code{\link[base:mean]{base::mean()}} and
\code{\link[base:Log]{base::exp()}}.
}
