% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{get_cookie_value}
\alias{get_cookie_value}
\alias{check_cookie_value}
\title{Extract cookie value by name}
\usage{
get_cookie_value(cookie_header, cookie_name)

check_cookie_value(cookie_value)
}
\arguments{
\item{cookie_header}{A raw Cookie header string (e.g. "a=1; b=2").}

\item{cookie_name}{A single character string with a cookie name to extract.}

\item{cookie_value}{A single character string with a cookie value.}
}
\value{
\itemize{
\item \code{get_cookie_value()} returns a single character string with the cookie
value or NA if not found.
\item \code{check_cookie_value()} returns the cookie value or "unknown" if empty.
}
}
\description{
Gets a specific cookie value from a raw HTTP Cookie header.
}
