% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{get_decade_aggregator}
\alias{get_decade_aggregator}
\alias{add_decade_column}
\alias{set_decade_aggregator}
\alias{run_decade_aggregator}
\title{Create and apply an aggregation strategy}
\usage{
get_decade_aggregator(aggregation_params)

add_decade_column(data_table)

set_decade_aggregator(aggregation_params)

run_decade_aggregator(aggregation_strategy, data_list)
}
\arguments{
\item{aggregation_params}{A list containing the function name that will be
used for the aggregation and the column name on which the calculation will
be performed.}

\item{data_table}{A \code{data.table} object to which aggregation will be applied.}

\item{aggregation_strategy}{A function which is a relevant aggregation
strategy.}

\item{data_list}{A list on which aggregation strategy will be used. It should
contain data objects that can be coerced to \code{data.table}.}
}
\value{
See Details.
}
\description{
Prepares and applies the aggregation strategy that transforms the annual
country indicator data into decades.
}
\details{
\strong{\code{get_decade_aggregator}} utilizes \code{aggregation_params} to create the
aggregation strategy (function) which will convert the annual country
indicator data into a ten-year summary. The aggregation function needs a
\code{data.table} object as an input.

\strong{\code{add_decade_column}} adds "Decade" column to a data table. Both input and
output are \code{data.table} objects.

\strong{\code{set_decade_aggregator}} prepares function that will apply given
aggregation strategy to a data object that can be coerced to \code{data.table}.

\strong{\code{run_decade_aggregator}} applies the aggregation strategy to the list
of data tables, merges them together and reshapes from long to wide format.
Returns a single \code{data.table} with the aggregated data per country and
decade.
}
