% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_indicator_selector.R
\name{indicatorSelectorUI}
\alias{indicatorSelectorUI}
\alias{indicatorSelectorServer}
\title{Indicator selector Shiny module}
\usage{
indicatorSelectorUI(id)

indicatorSelectorServer(id, choices = set_name_id_indicators())
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{choices}{A named vector or list with a choices for the selector.
Passed to \code{\link[shiny:selectInput]{shiny::selectInput()}} as \code{choices} parameter. Default
values are provided by \code{\link[=set_name_id_indicators]{set_name_id_indicators()}}.}
}
\value{
A reactive expression with a selected value inside.
}
\description{
Provides UI and server logic for an indicator selector.
}
