% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{make_chart_color_theme}
\alias{make_chart_color_theme}
\title{Make chart color theme}
\usage{
make_chart_color_theme(
  series_color,
  legend_color = set_color_palette()$transparent
)
}
\arguments{
\item{series_color}{A string that defines the trend chart series line
and marker colors.}

\item{legend_color}{A string that defines the trend chart legend
background and border colors. Default value is provided by
\code{\link[=set_color_palette]{set_color_palette()}}.}
}
\value{
A named list with a color theme for the trend chart series
and legend.
}
\description{
Makes a color theme to use in the trends chart series and legend.
}
