% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{make_chart_config}
\alias{make_chart_config}
\title{Make chart configuration}
\usage{
make_chart_config(
  indic_id,
  initial_range,
  indic_name = names(which(set_name_id_indicators() == indic_id))
)
}
\arguments{
\item{indic_id}{A string with a data indicator id.}

\item{initial_range}{A vector containing a pair of integers where
the first element is the data range real minimum
and the second is the real maximum.}

\item{indic_name}{A string with a data indicator "human friendly" name.
Default values are extracted from \code{\link[=set_name_id_indicators]{set_name_id_indicators()}}.}
}
\value{
A named list with a configuration for the trend chart layout.
}
\description{
Makes a configuration list to use in the trends chart layout.
}
