% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{make_decade_aggregator}
\alias{make_decade_aggregator}
\alias{run_aggregation_pipeline}
\title{Make and apply an aggregation function}
\usage{
make_decade_aggregator(aggregation_params)

run_aggregation_pipeline(aggregation_strategy, data_list)
}
\arguments{
\item{aggregation_params}{A list containing the function name that will be
used for the aggregation and the column name on which the calculation will
be performed.}

\item{aggregation_strategy}{A function which is a relevant aggregation
strategy.}

\item{data_list}{A list on which aggregation strategy will be used. It should
contain \code{data.table} objects.}
}
\value{
See Details.
}
\description{
Prepares and applies the aggregation strategy that transforms the annual
country indicator data into decades.
}
\details{
\strong{\code{make_decade_aggregator}} utilizes \code{aggregation_params} to make the
aggregation strategy (function) which will convert the annual country
indicator data into a ten-year summary. The aggregation function needs an
object that can be coerced to \code{data.table} as an input.
\strong{\code{run_aggregation_pipeline}} applies the aggregation strategy to the list
of data tables, merges them together and reshapes from long to wide format.
Returns a single \code{data.table} with the aggregated data per country and
decade.
}
