% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_trend_chart.R
\name{make_download_file_name}
\alias{make_download_file_name}
\title{Make trends chart download file name}
\usage{
make_download_file_name(
  name_root,
  name_prefix = format(Sys.Date(), "\%Y\%m\%d")
)
}
\arguments{
\item{name_root}{A string containing the root of the file name.
I.e. World Bank data indicator name.}

\item{name_prefix}{A string containing the prefix of the file name.
I.e. current system date in "YYYYMMDD" format.}
}
\value{
A string that is the file name.
}
\description{
Creates a name for a download file that contains the trend chart data.
}
\examples{
make_download_file_name("Unemployment (\%, 15 yo and over, modeled)")
}
