% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_world_map.R
\name{make_map_color_theme}
\alias{make_map_color_theme}
\title{Make map color theme}
\usage{
make_map_color_theme(
  color_palette = set_color_palette(),
  target_names = c("stroke", "fill", "highlight_stroke")
)
}
\arguments{
\item{color_palette}{A named list of colors used in the app. Default values
are provided by \code{\link[=set_color_palette]{set_color_palette()}}}

\item{target_names}{A character vector with element names for the color
theme.}
}
\value{
A named list with a color theme for "clickable" and "notclickable"
map polygons.
}
\description{
Makes a color theme to use in the world map polygons.
}
