% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_values_manager.R
\name{make_sql_qualified_columns}
\alias{make_sql_qualified_columns}
\title{Make SQL table column names in qualified format}
\usage{
make_sql_qualified_columns(db_schema)
}
\arguments{
\item{db_schema}{A named list with SQL table names as the list element names
and column names as the list elements (character vectors).}
}
\value{
A named list with SQL table names as the list element names and
column names, in a qualified column name format, as the list elements
(character vectors).
}
\description{
Converts SQL table column names to a qualified column name format.
}
\examples{
new_db_schema <- list(
  table_name1 = c("col1", "col2", "col3"),
  table_name2 = c("col1", "col4")
)
make_sql_qualified_columns(new_db_schema)
}
