% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myportfolio-package.R
\docType{package}
\name{myportfolio-package}
\alias{myportfolio}
\alias{myportfolio-package}
\title{myportfolio: My Portfolio App}
\description{
Functions and classes to make the app run easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cieslarjacek/myportfolio}
}

}
\author{
\strong{Maintainer}: Jacek Cieslar \email{contact@jacekdev.com}

}
\keyword{internal}
