% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_values_manager.R
\name{ReactiveValuesManager}
\alias{ReactiveValuesManager}
\title{Reactive values objects initialization manager}
\description{
Initiates reactive values predefined in \code{get_reactive_values_init()}.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{init_objects}}{A read only list of \code{reactiveValues}
and \code{reactiveVal} objects in their initial state.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReactiveValuesManager-new}{\code{ReactiveValuesManager$new()}}
\item \href{#method-ReactiveValuesManager-get_init}{\code{ReactiveValuesManager$get_init()}}
\item \href{#method-ReactiveValuesManager-update}{\code{ReactiveValuesManager$update()}}
\item \href{#method-ReactiveValuesManager-reset}{\code{ReactiveValuesManager$reset()}}
\item \href{#method-ReactiveValuesManager-clone}{\code{ReactiveValuesManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValuesManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValuesManager-new}{}}}
\subsection{Method \code{new()}}{
Creates a new manager object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValuesManager$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ReactiveValuesManager} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValuesManager-get_init"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValuesManager-get_init}{}}}
\subsection{Method \code{get_init()}}{
Retrieves target reactive object initial value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValuesManager$get_init(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A string that is a name in the \code{.init_objects} list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{reactiveValues} or \code{reactiveVal} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValuesManager-update"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValuesManager-update}{}}}
\subsection{Method \code{update()}}{
Updates target reactive object value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValuesManager$update(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A string that is a name in the \code{.init_objects} list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A function responsible for updating \code{reactiveValues}
or \code{reactiveVal} object value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValuesManager-reset"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValuesManager-reset}{}}}
\subsection{Method \code{reset()}}{
Reset target reactive object value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValuesManager$reset(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{A string that is a name in the \code{.init_objects} list.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An initial value for \code{reactiveValues} or \code{reactiveVal} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReactiveValuesManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReactiveValuesManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReactiveValuesManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
