% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_values_manager.R
\name{remove_list_elem}
\alias{remove_list_elem}
\title{Remove a list element by name}
\usage{
remove_list_elem(list_object, elem_name)
}
\arguments{
\item{list_object}{A named list.}

\item{elem_name}{A name of the element to remove.}
}
\value{
A named list with removed element.
}
\description{
Removes an element from a named list. It is mainly used as a helper
for updating \code{reactiveVal} objects that hold a list inside.
}
