% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sapply_identical}
\alias{sapply_identical}
\title{Find a list element that is identical to the given value}
\usage{
sapply_identical(target_list, expected_value)
}
\arguments{
\item{target_list}{A list of actual elements to check.}

\item{expected_value}{An expected value. Can be any R object.}
}
\value{
A vector of logical values.
}
\description{
Checks if actual list contains element that is identical to the expected
value.
}
