% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_reader.R
\name{SecretReader}
\alias{SecretReader}
\title{Secrets reader}
\description{
Reads secrets from the vault.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{current_env}}{The environment where the app is deployed and runs.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SecretReader-new}{\code{SecretReader$new()}}
\item \href{#method-SecretReader-get_all_secrets}{\code{SecretReader$get_all_secrets()}}
\item \href{#method-SecretReader-get_secret}{\code{SecretReader$get_secret()}}
\item \href{#method-SecretReader-clone}{\code{SecretReader$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecretReader-new"></a>}}
\if{latex}{\out{\hypertarget{method-SecretReader-new}{}}}
\subsection{Method \code{new()}}{
Creates a new reader object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecretReader$new(current_env = Sys.getenv("DEPLOYMENT_ENV"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{current_env}}{A string containing the name of the environment where
the app is deployed and runs.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{SecretReader} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecretReader-get_all_secrets"></a>}}
\if{latex}{\out{\hypertarget{method-SecretReader-get_all_secrets}{}}}
\subsection{Method \code{get_all_secrets()}}{
Retrieves all secrets from the target vault.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecretReader$get_all_secrets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of secrets for the current environment.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecretReader-get_secret"></a>}}
\if{latex}{\out{\hypertarget{method-SecretReader-get_secret}{}}}
\subsection{Method \code{get_secret()}}{
Retrieves a single secret value from the target vault.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecretReader$get_secret(secret_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{secret_name}}{A string containing the name of the secret.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A secret value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SecretReader-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SecretReader-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SecretReader$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
