% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sidenote.R
\name{sidenoteUI}
\alias{sidenoteUI}
\alias{sidenoteServer}
\alias{get_sidenote_message}
\title{Side note Shiny module}
\usage{
sidenoteUI(id)

sidenoteServer(id, message = get_sidenote_message(id))

get_sidenote_message(id)
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{message}{A character vector with a message to show. Default value
is provided by \code{\link[=get_sidenote_message]{get_sidenote_message()}}.}
}
\description{
Provides UI and server logic for creating notes inside side panels.
}
