% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_components.R
\name{source_components}
\alias{source_components}
\title{Source component files}
\usage{
source_components(component_paths, evalutation_env = globalenv())
}
\arguments{
\item{component_paths}{A character vector with component file paths.}

\item{evalutation_env}{Environment in which the parsed expressions are
evaluated. Passed to \code{\link[base:source]{base::source()}} as \code{local} parameter. Default is
\code{globalenv()}.}
}
\description{
Sources each file in \code{component_paths} set using \code{\link[base:source]{base::source()}}.
}
