% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{source_weight}
\alias{source_weight}
\title{Source a weight data from one data table to another}
\usage{
source_weight(data_pair)
}
\arguments{
\item{data_pair}{A list containing pair of indicator data (see Details).}
}
\value{
The "receiver" list containing the same number of elements, but with
added column "weight" to each.
}
\description{
Extracts "value" column as "weight" from one indicator data table to another
using "country_id" and "time_period" as merge controls.
}
\details{
\strong{\code{data_pair}} list needs to have exactly two elements. Both elements have
to be a named lists with the indicator data tables per country. Data tables
have to contain the same columns. The second element is always used as
a "giver" and the first one is always "receiver".
}
