% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_service.R
\name{StorageService}
\alias{StorageService}
\title{Storage service}
\description{
Manages writing to and fetching from dedicated MinIO storage.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StorageService-new}{\code{StorageService$new()}}
\item \href{#method-StorageService-get_data}{\code{StorageService$get_data()}}
\item \href{#method-StorageService-get_geometry_all}{\code{StorageService$get_geometry_all()}}
\item \href{#method-StorageService-put_data}{\code{StorageService$put_data()}}
\item \href{#method-StorageService-put_geometry_all}{\code{StorageService$put_geometry_all()}}
\item \href{#method-StorageService-clone}{\code{StorageService$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-new"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-new}{}}}
\subsection{Method \code{new()}}{
Creates a new service object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$new(con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{A connection details for MinIO storage. Please check
\code{\link[aws.s3:s3HTTP]{aws.s3::s3HTTP()}} for more details.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{StorageService} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-get_data}{}}}
\subsection{Method \code{get_data()}}{
Fetches relevant data from the target storage bucket.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$get_data(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string indicating type of the data to fetch.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-get_geometry_all"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-get_geometry_all}{}}}
\subsection{Method \code{get_geometry_all()}}{
Fetches geometry data (for all countries) from the storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$get_geometry_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-put_data"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-put_data}{}}}
\subsection{Method \code{put_data()}}{
Writes relevant data to the target storage bucket.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$put_data(type, source_object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{A string indicating type of the data to write.}

\item{\code{source_object}}{A single object to be saved in the storage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-put_geometry_all"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-put_geometry_all}{}}}
\subsection{Method \code{put_geometry_all()}}{
Writes geometry data (for all countries) to the storage.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$put_geometry_all(source_object)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{source_object}}{A single object to be saved in the storage.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StorageService-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StorageService-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StorageService$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
