% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trend_chart.R
\name{trendChartExportBtnUI}
\alias{trendChartExportBtnUI}
\alias{trendChartExportBtnServer}
\title{Trends chart export button Shiny module}
\usage{
trendChartExportBtnUI(id)

trendChartExportBtnServer(id, export_types = set_export_types())
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{export_types}{A named character vector where the element names are
data export format types (e.g. png, csv, json) and the elements are
the button/link labels (e.g. "CSV data"),}
}
\description{
Provides UI and server logic for the data export button for the trends chart.
}
