% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trend_chart.R
\name{trendChartUI}
\alias{trendChartUI}
\alias{trendChartServer}
\alias{updtTrendChartYaxisServer}
\alias{addTrendChartSeriesServer}
\alias{removeTrendChartSeriesServer}
\title{Trends chart Shiny module}
\usage{
trendChartUI(id)

trendChartServer(id, render_flag, base_chart)

updtTrendChartYaxisServer(id, new_yrange)

addTrendChartSeriesServer(id, new_series, new_yrange)

removeTrendChartSeriesServer(id, series_label, new_yrange)
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{render_flag}{A \code{reactiveVal} object containing Boolean value.}

\item{base_chart}{An HTML Widget object containing \code{plotly} chart.}

\item{new_yrange}{A pair of integers - current real (no extra padding)
minimum and maximum values of y-axis.}

\item{new_series}{A two element vector containing new series data and color
(see Details).}

\item{series_label}{A string containing the name of one of the data series
already present on the trend chart.}
}
\description{
Provides UI and server logic for the \code{plotly} class trends chart.
.
}
\details{
\strong{\code{new_series}} vector needs to have two elements:
\itemize{
\item "dt_list" - a single-element, named list that contains a data table with
a values to plot. The name of the element is a country label (name).
\item "color" - a string with a line and marker color for the trend chart
series.
}
}
