% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_values_manager.R
\name{update_aggregation_specs}
\alias{update_aggregation_specs}
\title{Update \code{table_aggregation_params()}}
\usage{
update_aggregation_specs(
  indic_id,
  aggregation_params = set_id_aggr_indicators()[[indic_id]]
)
}
\arguments{
\item{indic_id}{A string with a data indicator id.}

\item{aggregation_params}{A named list with the aggregation function name,
target column name and optionally weight data source (database table name).
Default values are provide by \code{\link[=set_id_aggr_indicators]{set_id_aggr_indicators()}}.}
}
\value{
A named list containing aggregation parameters.
}
\description{
Selects new set of aggregation parameters for \code{table_aggregation_params()}
base on the indicator id.
}
