% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactive_values_manager.R
\name{update_db_schema}
\alias{update_db_schema}
\title{Update \code{sql_db_schema()}}
\usage{
update_db_schema(
  indic_id,
  weight_id = set_id_aggr_indicators()[[indic_id]]$weight_source,
  db_schema = make_sql_qualified_columns(set_db_schema_registry()$world_trends)
)
}
\arguments{
\item{indic_id}{A string with a data indicator id.}

\item{weight_id}{A string with a data weight id. If \code{NULL} no additional data
are extracted in order to perform the indicator data aggregation. Default
value is provided by \code{\link[=set_id_aggr_indicators]{set_id_aggr_indicators()}}.}

\item{db_schema}{A named list of length four. It contains SQL table names as
the list element names and column names, in a qualified column name
format, as the list elements (character vectors).
Please check \code{\link[=make_sql_qualified_columns]{make_sql_qualified_columns()}} and \code{\link[=set_db_schema_registry]{set_db_schema_registry()}}
for details about the default value.}
}
\value{
An updated version of \code{db_schema} with replaced strings in
the list names and related elements. It can be one element shorter than
initial list if \code{weight_id} was \code{NULL}.
}
\description{
Substitutes old table names in \code{sql_db_schema()} with a new ones. The names
pinpoint to the database tables related to the selected data indicator.
}
\examples{
\dontrun{
old_db_schema <- list(
  table_name1 = c(
    "table_name1.col1", "table_name1.col2", "table_name1.col3"
  ),
  table_name2 = c(
    "table_name2.col1", "table_name2.col4"
  ),
  table_name3 = c(
    "table_name3.col1", "table_name3.col2", "table_name3.col3"
  ),
  table_name4 = c(
    "table_name4.col7"
  )
)
update_db_schema("table_name99", NULL, old_db_schema)
}
}
