% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_world_map.R
\name{worldMapUI}
\alias{worldMapUI}
\alias{worldMapServer}
\title{World map Shiny module}
\usage{
worldMapUI(id)

worldMapServer(id, indic_id, data_pair)
}
\arguments{
\item{id}{Module ID. Check \code{\link[shiny:NS]{shiny::NS()}} for more details.}

\item{indic_id}{A \code{reactive} object containing string with a data indicator
id.}

\item{data_pair}{A list containing pair of country data - geometry (a list
of \code{sf} data frames) and labels.}
}
\description{
Provides UI and server logic for a \code{leaflet} class world map.
}
