% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{add_default_spinner}
\alias{add_default_spinner}
\title{Add a spinner}
\usage{
add_default_spinner(ui_element, spinner_color = set_color_palette()$primary_lt)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a spinner
when the corresponding output is being calculated. Passed to
\code{\link[shinycssloaders:withSpinner]{shinycssloaders::withSpinner()}} as \code{ui_element} parameter.}

\item{spinner_color}{A string that defines the spinner color. Default value
is provided by \code{\link[=set_color_palette]{set_color_palette()}}.}
}
\description{
Utilizes \code{\link[shinycssloaders:withSpinner]{shinycssloaders::withSpinner()}} with a default settings.
}
