% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{add_js_code}
\alias{add_js_code}
\alias{get_js_event}
\title{Add JavaScript code to HTML Widget object}
\usage{
add_js_code(htmlwidget_object, id, js_code = get_js_event(id))

get_js_event(id)
}
\arguments{
\item{htmlwidget_object}{An HTML Widget object to which JavaScript code
snippet will be attached. Please check \link[htmlwidgets:htmlwidgets-package]{htmlwidgets::htmlwidgets} for more details.}

\item{id}{\code{htmlwidget_object} ID.}

\item{js_code}{A character vector containing JavaScript code. Please check
\code{\link[htmlwidgets:onRender]{htmlwidgets::onRender()}} for more details. Default value is provided by
\code{\link[=get_js_event]{get_js_event()}}.}
}
\value{
The modified widget object.
}
\description{
Attaches custom JavaScript code to HTML Widget object (e.g. map, chart,
table).
}
