% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_settings.R
\name{get_app_settings}
\alias{get_app_settings}
\alias{set_app_components}
\alias{set_color_palette}
\alias{set_data_labels}
\alias{set_db_schema_registry}
\alias{set_export_types}
\alias{set_name_id_indicators}
\alias{set_id_aggr_indicators}
\alias{set_aggr_config}
\title{Get list of predefined app settings}
\usage{
get_app_settings()

set_app_components()

set_color_palette()

set_data_labels()

set_db_schema_registry()

set_export_types()

set_name_id_indicators()

set_id_aggr_indicators()

set_aggr_config(
  ...,
  target_names = c("function_name", "column_name", "weight_source"),
  length_range = setNames(c(2, 3), c("min", "max"))
)
}
\arguments{
\item{...}{A character vector of length two or three containing elements for
aggregation parameters list.}

\item{target_names}{A character vector containing element names for
aggregation parameters list - the aggregation function name, target
column name and optionally weight data source (database table name).
Default is \code{c("function_name", "column_name", "weight_source")}.}

\item{length_range}{A named integer vector containing information about
minimum and maximum number of aggregation parameters.
Default is \code{setNames(c(2, 3), c("min", "max"))}.}
}
\description{
Get list of predefined app settings
}
\details{
\code{get_app_settings()} returns a named list containing a multiple elements with
a data that are used across the whole app. Each element is created by the
related "set_" function.
\itemize{
\item "app_component" is a character vector of component sub-directories.
\item "color_palette" is a named list of colors.
\item "data_labels" is a character vector of selected data labels.
\item "db_schema_registry" is a named list with SQL table names as
the list element names and column names as the list elements
(character vectors).
\item "export_types" is a named character vector where the element names are
data export format types (e.g. png, csv, json) and the elements are related
labels (e.g. "CSV data").
\item "worldbank_indicator$name_id" is a named list where the elements are
data indicator ids and the element names are indicator "human friendly"
names.
\item "worldbank_indicator$id_aggr" is a named list where the elements are
the data indicator aggregation parameters (see \code{target_names} description)
and the element names are data indicator ids.
}
}
