% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_service.R
\name{get_storage_connection}
\alias{get_storage_connection}
\title{Get a connection details for MinIO storage}
\usage{
get_storage_connection(secrets)
}
\arguments{
\item{secrets}{A named character vector that contains a configuration data
for MinIO storage connection.}
}
\value{
A named list with connection details for MinIO storage.
}
\description{
Get a connection details for relevant MinIO storage from provided secrets.
}
