% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_page.R
\name{get_ui_pattern}
\alias{get_ui_pattern}
\alias{create_ui_pattern}
\title{Get or create \code{uiPattern}}
\usage{
get_ui_pattern()

create_ui_pattern(url_path)
}
\arguments{
\item{url_path}{A character vector with target URL path(s). URL path
should start with "/".}
}
\value{
A regex string.
}
\description{
Gets general \code{uiPattern} for \code{\link[shiny:shinyApp]{shiny::shinyApp()}} that will match every
URL path that starts with "/" or creates regex with exact \code{uiPattern}.
}
