% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_empty}
\alias{is_empty}
\title{Check if object is empty}
\usage{
is_empty(target_object)
}
\arguments{
\item{target_object}{An object/value to be tested.}
}
\value{
A single logical value.
}
\description{
Invalidates provided object.
}
\details{
It is a modified application of \code{invalid()} from \pkg{gtools} -
\url{https://github.com/r-gregmisc/gtools/blob/master/R/invalid.R}.
}
