% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route_page.R
\name{route_page}
\alias{route_page}
\alias{select_page}
\title{Route page}
\usage{
route_page(route_schema, default = "404 Not Found")

select_page(url_path, route_schema, default)
}
\arguments{
\item{route_schema}{A list of URL paths to target pages (UI objects) mapping.}

\item{default}{A page (UI object) to which all invalid URL paths
are redirected. Default is "404 Not Found".}

\item{url_path}{A string with the URL path.}
}
\value{
\itemize{
\item \code{route_page()} returns a function that accepts a Rook environment
(\code{rook_env}) and, when called, returns the matched page.
\item \code{select_page()} returns the matched page: a \code{shiny.tag} or
\code{shiny.tag.list} class object if \code{url_path} matches an entry in
\code{route_schema}, or \code{default} otherwise.
}
}
\description{
Routes to expected page (UI object) using the Rook environment properties.
}
\details{
Page (UI object) is selected based on matching URL path presented in the Rook
environment.
}
