% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_shiny.R
\name{run_validate_need}
\alias{run_validate_need}
\alias{set_validation_message}
\title{Just a wrapper for a classic Shiny \code{validate(need())} combo}
\usage{
run_validate_need(
  expression,
  output_id = shiny::getCurrentOutputInfo()$name,
  message = set_validation_message(output_id)
)

set_validation_message(output_id)
}
\arguments{
\item{expression}{An expression to test.}

\item{output_id}{A string containing output id/name in which the validation
is run. Default value is obtained using \code{\link[shiny:getCurrentOutputInfo]{shiny::getCurrentOutputInfo()}}.}

\item{message}{A message to show to the user if the test fails. Default value
is provided by \code{\link[=set_validation_message]{set_validation_message()}}.}
}
\description{
Runs \code{\link[shiny:validate]{shiny::validate()}} with a single \code{\link[shiny:validate]{shiny::need()}}.
}
