% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_dt.R
\name{select_in_dt}
\alias{select_in_dt}
\title{Filter data table rows using \code{\%in\%} operator}
\usage{
select_in_dt(data_object, col_name, target_values)
}
\arguments{
\item{data_object}{An object that can be coerced to \code{data.table}.}

\item{col_name}{A character string with a column name for which data table
operation will be conducted.}

\item{target_values}{A vector containing expected values.}
}
\value{
A data table with selected rows.
}
\description{
Selects data table rows that are \code{\%in\%} the expected values for
the given column.
}
