% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_range.R
\name{select_range}
\alias{select_range}
\title{Select range for data}
\usage{
select_range(range_pair)
}
\arguments{
\item{range_pair}{A list containing two range elements. First element is the
current (user modified) data range and the second element is the default
(absolute) data range.}
}
\value{
A vector containing a pair of integers where the first element is
the data range minimum and the second is the maximum.
}
\description{
Selects range for data from two available choices.
}
\details{
From \code{range_pair} the "current" one is always preferred over the "default".
}
