% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_logger.R
\name{SessionLogger}
\alias{SessionLogger}
\title{Session Logger}
\description{
Captures all console output (stdout and messages) for a Shiny session in a
local file. It can be used to control when to start and stop logging,
and when to write logs to MinIO storage.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{time_stamp}}{A character string with current date and time in
"\%Y\%m\%d_\%H\%M\%S" format.}

\item{\code{log_file_dir}}{A name of the directory for storing log files.}

\item{\code{log_file_name}}{A log file name in "\%Y\%m\%d_\%H\%M\%S.log" format.}

\item{\code{log_file_path}}{A path to the log file.}

\item{\code{log_file_con}}{A connection to the log file.}

\item{\code{current_session_token}}{A named list of currently active session tokens
for the given visit ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionLogger-new}{\code{SessionLogger$new()}}
\item \href{#method-SessionLogger-start}{\code{SessionLogger$start()}}
\item \href{#method-SessionLogger-stop}{\code{SessionLogger$stop()}}
\item \href{#method-SessionLogger-put}{\code{SessionLogger$put()}}
\item \href{#method-SessionLogger-get_visit_id}{\code{SessionLogger$get_visit_id()}}
\item \href{#method-SessionLogger-set_session_token}{\code{SessionLogger$set_session_token()}}
\item \href{#method-SessionLogger-get_session_token}{\code{SessionLogger$get_session_token()}}
\item \href{#method-SessionLogger-log_session_event}{\code{SessionLogger$log_session_event()}}
\item \href{#method-SessionLogger-init_session}{\code{SessionLogger$init_session()}}
\item \href{#method-SessionLogger-terminate_session}{\code{SessionLogger$terminate_session()}}
\item \href{#method-SessionLogger-clone}{\code{SessionLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{SessionLogger} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{SessionLogger} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-start"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-start}{}}}
\subsection{Method \code{start()}}{
Starts capturing console output for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-stop"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-stop}{}}}
\subsection{Method \code{stop()}}{
Stops capturing console output for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-put"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-put}{}}}
\subsection{Method \code{put()}}{
Reads and writes relevant log file to the target storage bucket.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$put(storage_con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage_con}}{A connection details for MinIO storage. Please check
\code{\link[aws.s3:s3HTTP]{aws.s3::s3HTTP()}} for more details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-get_visit_id"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-get_visit_id}{}}}
\subsection{Method \code{get_visit_id()}}{
Returns the value of \code{visit_id} cookie.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$get_visit_id(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-set_session_token"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-set_session_token}{}}}
\subsection{Method \code{set_session_token()}}{
Register the currently active session token for a given \code{visit_id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$set_session_token(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-get_session_token"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-get_session_token}{}}}
\subsection{Method \code{get_session_token()}}{
Returns the currently active session token registered for a given
\code{visit_id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$get_session_token(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A single session token string or \code{NULL} if not found.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-log_session_event"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-log_session_event}{}}}
\subsection{Method \code{log_session_event()}}{
Writes a uniform session event line to the log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$log_session_event(session, event_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}

\item{\code{event_name}}{A character string with the relevant event name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-init_session"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-init_session}{}}}
\subsection{Method \code{init_session()}}{
Initializes user session:
\enumerate{
\item Register the currently active session token for a given \code{visit_id}.
\item Log a session start event with timestamp, \code{PATH_INFO}, \code{visit_id}
cookie value and session token.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$init_session(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-terminate_session"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-terminate_session}{}}}
\subsection{Method \code{terminate_session()}}{
Terminates users session:
\enumerate{
\item Grab the currently active session token registered for a given
\code{visit_id}.
\item If it's correct log a session end event with timestamp, \code{PATH_INFO},
\code{visit_id} cookie value and session token.
\item Write the log file to the storage.
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$terminate_session(session, con_details)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}

\item{\code{con_details}}{A connection details for the storage. Please check
\code{put} method in \link{SessionLogger} for more details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
