% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_logger.R
\name{SessionLogger}
\alias{SessionLogger}
\title{Session Logger}
\description{
Captures all console output (stdout and messages) for a Shiny session in a
local file. It can be used to control when to start and stop logging,
and when to write logs to MinIO storage.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{time_stamp}}{A character string with current date and time in
"\%Y\%m\%d_\%H\%M\%S" format.}

\item{\code{log_file_dir}}{A name of the directory for storing log files.}

\item{\code{log_file_name}}{A log file name in "\%Y\%m\%d_\%H\%M\%S.log" format.}

\item{\code{log_file_path}}{A path to the log file.}

\item{\code{log_file_con}}{A connection to the log file.}

\item{\code{current_session_token}}{A currently active session token for the given
visit ID.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SessionLogger-new}{\code{SessionLogger$new()}}
\item \href{#method-SessionLogger-start}{\code{SessionLogger$start()}}
\item \href{#method-SessionLogger-stop}{\code{SessionLogger$stop()}}
\item \href{#method-SessionLogger-put}{\code{SessionLogger$put()}}
\item \href{#method-SessionLogger-log_session_event}{\code{SessionLogger$log_session_event()}}
\item \href{#method-SessionLogger-log_session_start}{\code{SessionLogger$log_session_start()}}
\item \href{#method-SessionLogger-log_session_end}{\code{SessionLogger$log_session_end()}}
\item \href{#method-SessionLogger-clone}{\code{SessionLogger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-new"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{SessionLogger} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{SessionLogger} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-start"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-start}{}}}
\subsection{Method \code{start()}}{
Starts capturing console output for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-stop"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-stop}{}}}
\subsection{Method \code{stop()}}{
Stops capturing console output for this session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-put"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-put}{}}}
\subsection{Method \code{put()}}{
Reads and writes relevant log file to the target storage bucket.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$put(storage_con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{storage_con}}{A connection details for MinIO storage. Please check
\code{\link[aws.s3:s3HTTP]{aws.s3::s3HTTP()}} for more details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-log_session_event"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-log_session_event}{}}}
\subsection{Method \code{log_session_event()}}{
Writes a uniform session event line to the log.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$log_session_event(session, event_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}

\item{\code{event_name}}{A character string with the relevant event name.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-log_session_start"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-log_session_start}{}}}
\subsection{Method \code{log_session_start()}}{
Logs a session start event with timestamp, \code{PATH_INFO}, \code{visit_id}
cookie value and session token.
Registers this session as the currently active sub-session for its
\code{visit_id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$log_session_start(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-log_session_end"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-log_session_end}{}}}
\subsection{Method \code{log_session_end()}}{
Logs a session end event with timestamp, \code{PATH_INFO}, \code{visit_id}
cookie value and session token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$log_session_end(session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session}}{A Shiny session object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SessionLogger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SessionLogger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SessionLogger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
